Add GoalCheeringEntity with DTO
