+++ raisedragon-core/src/main/kotlin/com/whatever/raisedragon/domain/goalcheering/GoalCheering.kt
@@ -1,26 +1,21 @@
 package com.whatever.raisedragon.domain.goalcheering
 
-import com.whatever.raisedragon.domain.BaseEntity
-import com.whatever.raisedragon.domain.goal.GoalEntity
-import com.whatever.raisedragon.domain.user.UserEntity
-import jakarta.persistence.*
+import java.time.LocalDateTime
 
-@Table(name = "goal_cheering")
-@Entity
-class GoalCheering(
+data class GoalCheering(
+    val id: Long,
+    val userId: Long,
+    val cheeringMessage: CheeringMessage,
+    val createdAt: LocalDateTime,
+    val updatedAt: LocalDateTime,
+    val deletedAt: LocalDateTime?
+)
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "goal_id")
-    val goalEntity: GoalEntity,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    val user: UserEntity,
-
-    @Embedded
-    val cheeringMessage: CheeringMessage
-
-) : BaseEntity()
-
-@Embeddable
-data class CheeringMessage(val cheeringMessage: String)
+fun GoalCheeringEntity.toDTO(): GoalCheering = GoalCheering(
+    id = id,
+    userId = user.id,
+    cheeringMessage = cheeringMessage,
+    createdAt = createdAt,
+    updatedAt = updatedAt,
+    deletedAt = deletedAt
+)
